docs: add minor fixes after review
This PR aims to fix a few minor topics after
a review from Foundation for Public Code.
It is part of an ongoing process to improve the project.

Aims to fix the following doc content:
- Closer explain the connection to Digg's Internal Open Source Policy
and how they are connected. (This project can be seen as a practical
appendix from that.
- Adds a "Public known users list"
- Adds a note in Contributing that maintanance and review costs might at
  times have to be agreed upon, and gives some examples for when that
might occur.
- Adds a note about expected response times.

Structure:
- The SECURITY.md template was moved to SECURITY.template.md,
and a real SECURITY, based on that for this project was added.
- Same with CONTRIBUTING, a template version was added.
- The Standard moved to the CONTRIBUTING file

Signed-off-by: Josef Andersson <josef.andersson@digg.se>