feat: add project quality script
This PR adds a bash script that runs the same
checks that are run in the CI (linting, commits, license).
This makes it possible for a contributor to locally verify
the contribution before opening a PR, and so,
avoid any CI check surprises later on.

It adds a brief DEVELOPMENT file to explain how to run it.

It also fixed a few minor things found while running the
check.

Signed-off-by: Josef Andersson <josef.andersson@digg.se>